pyproject.toml
[project]
name = "project-manutenzione_preventiva_per_la_robotica_e_la_automazione_intelligente"
version = "0.1.0"
description = "..."
authors = [
    {name = "Catalinaqi",email = "ccequispe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "joblib (>=1.5.3,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = true
packages = [
    { include = "phm_america_2024", from = "src" }
]